move fourier_operations into algo.__init__
move tests into test_algo.py and remove tests for removed files
